ICS Deploy Update
When running in ICS we'll need to use client certificates to talk to ICS.
These certs are stored encrypted on a volume mapped into the service container at /data
A utility script lives in the container that can decrypt the certs and configure the environment vars for etcdctl to function as expected.
We can identify when we need to run this script, by the presence of the the ETCD_SECRET environment var, which carries the password required to decrypt the certificates. When this is set, we'll launch the utility script, and let it configure our environment.